refactor(cli): extract normalize_path to shared path_util module
The `normalize_path` function (filesystem-free `.`/`..` resolution)
was duplicated between `commands/deps/path.rs` (binary crate) and
`dependency_classes.rs` (library crate), with a "keep in sync"
comment flagged in BT-2836. This was caused by the binary/library
split inside beamtalk-cli: the library could not reach into the
binary's private `commands::` tree.

Since the binary already links against the library, the correct fix
is to move the single implementation into the library's new
`path_util` module (`pub fn normalize_path`) and have both callers
import it from there. The duplicate tests are consolidated into
`path_util`'s own `#[cfg(test)]` block.
