+++ crates/beamtalk-cli/src/commands/deps/path.rs
@@ -12,6 +12,7 @@
 //! - Detects circular dependencies
 //! - Returns ebin paths for BEAM code path setup
 
+use beamtalk_cli::path_util::normalize_path;
 use beamtalk_core::compilation::DependencySource;
 use camino::{Utf8Path, Utf8PathBuf};
 use miette::{Context, IntoDiagnostic, Result};
@@ -279,50 +280,6 @@ pub(crate) fn canonicalize_dep_path(
 
     // Normalize the path by resolving `.` and `..` components
     normalize_path(&joined)
-}
-
-/// Normalize a path by resolving `.` and `..` components without filesystem access.
-///
-/// Unlike `std::fs::canonicalize`, this does not require the path to exist and
-/// does not resolve symlinks.
-fn normalize_path(path: &Utf8Path) -> Utf8PathBuf {
-    use camino::Utf8Component;
-
-    let mut components = Vec::new();
-    for component in path.components() {
-        match component {
-            Utf8Component::CurDir => {
-                // Skip `.`
-            }
-            Utf8Component::ParentDir => match components.last().copied() {
-                // Already at the filesystem root — an extra `..` is a no-op
-                // rather than something to pop or accumulate (BT-2836 review
-                // finding: popping `RootDir` here would turn `/foo/../..`
-                // into `.` instead of `/`).
-                Some(Utf8Component::RootDir) => {}
-                // Nothing to pop yet, or a run of leading `..`s in a
-                // relative path — accumulate.
-                None | Some(Utf8Component::ParentDir) => components.push(component),
-                // A real component precedes it — cancel the two out.
-                Some(_) => {
-                    components.pop();
-                }
-            },
-            _ => {
-                components.push(component);
-            }
-        }
-    }
-
-    if components.is_empty() {
-        return Utf8PathBuf::from(".");
-    }
-
-    let mut result = Utf8PathBuf::new();
-    for component in components {
-        result.push(component.as_str());
-    }
-    result
 }
 
 /// Compile a dependency and return a `ResolvedDependency`.
@@ -715,36 +672,6 @@ mod tests {
     use super::*;
     use std::fs;
     use tempfile::TempDir;
-
-    #[test]
-    fn test_normalize_path_resolves_parent() {
-        let path = Utf8PathBuf::from("/home/user/project/../dep");
-        assert_eq!(normalize_path(&path), Utf8PathBuf::from("/home/user/dep"));
-    }
-
-    #[test]
-    fn test_normalize_path_resolves_dot() {
-        let path = Utf8PathBuf::from("/home/user/./project");
-        assert_eq!(
-            normalize_path(&path),
-            Utf8PathBuf::from("/home/user/project")
-        );
-    }
-
-    #[test]
-    fn test_normalize_path_multiple_parents() {
-        let path = Utf8PathBuf::from("/home/user/project/../../dep");
-        assert_eq!(normalize_path(&path), Utf8PathBuf::from("/home/dep"));
-    }
-
-    #[test]
-    fn test_normalize_path_parent_at_root() {
-        // A `..` chain that consumes every real component and then hits
-        // root must not pop `RootDir` itself — `/foo/../..` is still `/`,
-        // not `.` (BT-2836 review finding).
-        let path = Utf8PathBuf::from("/foo/../..");
-        assert_eq!(normalize_path(&path), Utf8PathBuf::from("/"));
-    }
 
     #[test]
     fn test_canonicalize_dep_path() {
